+++ examples/cm_futures/market/query_index_price_constituents.py
@@ -6,5 +6,4 @@ from binance.lib.utils import config_logging
 config_logging(logging, logging.DEBUG)
 
 cm_futures_client = CMFutures()
-
-logging.info(cm_futures_client.pm_exchange_info())
+logging.info(cm_futures_client.query_index_price_constituents(symbol="BTCUSD"))